requirements.txt
web3>=6.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
cryptography>=41.0.0
aiohttp>=3.9.0
eth-account>=0.11.0
